ghostscript: extend fuzzing (#7821)
* ghostscript: extend fuzzing

- Add two new fuzzers
- One new fuzzer that explore multiple colorings
- One new fuzzer that focuses on PDF exploration
- Add dictionary to PDF fuzzer
- Add small sized PDF seed to PDF fuzzer
- Ensure max_len on multiple coloring fuzzing

* Add myself to auto_ccs

* Fix typo

* set indentation

* add options file

* Address review: split fuzzers and have dict + seed locally

* nit

* clarify prepending of single byte

* fix styling

* clarify comments in all_colors fuzzer

* Addess final review
