+++ projects/ghostscript/Dockerfile
@@ -22,7 +22,10 @@ RUN git clone --branch VER-2-12-1 --single-branch --depth 1 https://git.savannah
 RUN git clone --single-branch --depth 1 git://git.ghostscript.com/ghostpdl.git ghostpdl
 RUN mkdir ghostpdl/fuzz
 
+COPY pdf_seeds $SRC/pdf_seeds
+COPY dicts $SRC/dicts
+
 WORKDIR ghostpdl
 
-COPY gstoraster_fuzzer.cc $SRC/
+COPY *.cc *.options *.h $SRC/
 COPY build.sh $SRC/
