+++ projects/ghostscript/gstoraster_fuzzer_all_colors.options
@@ -1,0 +1,2 @@
+[libfuzzer]
+max_len = 16834
